+++ app/src/main/java/cn/wthee/pcrtool/ui/NavGraph.kt
@@ -41,38 +41,38 @@ object Navigation {
     const val CHARACTER_LIST = "characterList"
     const val CHARACTER_DETAIL = "characterDetail"
     const val CHARACTER_BASIC_INFO = "characterBasicInfo"
-    const val UNIT_ID = "unitId"
     const val EQUIP_LIST = "equipList"
-    const val EQUIP_ID = "equipId"
     const val EQUIP_DETAIL = "equipDetail"
     const val RANK_EQUIP = "rankEquip"
     const val RANK_COMPARE = "rankCompare"
-    const val MAX_RANK = "maxRank"
-    const val LEVEL = "level"
-    const val RARITY = "rarity"
-    const val UNIQUE_EQUIP_LEVEL = "uniqueEquipLevel"
     const val EQUIP_COUNT = "equipCount"
     const val EQUIP_MATERIAL = "equipMaterial"
-
-    //工具
     const val TOOL_LEADER = "toolLeader"
     const val TOOL_GACHA = "toolGacha"
     const val TOOL_EVENT = "toolEvent"
     const val TOOL_GUILD = "toolGuild"
     const val TOOL_CLAN = "toolClanBattle"
     const val TOOL_CLAN_BOSS_INFO = "toolClanBattleInfo"
-    const val TOOL_CLAN_BOSS_ID = "toolClanBattleID"
-    const val TOOL_CLAN_BOSS_INDEX = "toolClanBattleIndex"
     const val TOOL_PVP = "toolPvpSearch"
     const val TOOL_NEWS = "toolNews"
     const val TOOL_NEWS_DETAIL = "toolNewsDetail"
-    const val TOOL_NEWS_KEY = "toolNewsKey"
     const val MAIN_SETTINGS = "mainSettings"
     const val APP_NOTICE = "appNotice"
     const val TWEET = "tweet"
     const val COMIC = "comic"
-    const val COMIC_ID = "comicId"
     const val ALL_SKILL = "allSkill"
+
+    const val UNIT_ID = "unitId"
+    const val EQUIP_ID = "equipId"
+    const val MAX_RANK = "maxRank"
+    const val LEVEL = "level"
+    const val RARITY = "rarity"
+    const val UNIQUE_EQUIP_LEVEL = "uniqueEquipLevel"
+    const val COMIC_ID = "comicId"
+    const val TOOL_CLAN_BOSS_ID = "toolClanBattleID"
+    const val TOOL_CLAN_BOSS_INDEX = "toolClanBattleIndex"
+    const val TOOL_NEWS_KEY = "toolNewsKey"
+
 }
 
 @ExperimentalCoilApi
@@ -86,26 +86,8 @@ object Navigation {
 fun NavGraph(
     navController: NavHostController,
     viewModel: NavViewModel,
-    actions: NavActions,
-    newsViewModel: NewsViewModel = hiltViewModel()
+    actions: NavActions
 ) {
-
-    val lifecycle = LocalLifecycleOwner.current.lifecycle
-    newsViewModel.getNews(2)
-    newsViewModel.getNews(3)
-    newsViewModel.getNews(4)
-    val flow0 = newsViewModel.newsPageList0
-    val flow1 = newsViewModel.newsPageList1
-    val flow2 = newsViewModel.newsPageList2
-    val news0 = remember(flow0, lifecycle) {
-        flow0?.flowWithLifecycle(lifecycle = lifecycle)
-    }?.collectAsLazyPagingItems()
-    val news1 = remember(flow1, lifecycle) {
-        flow1?.flowWithLifecycle(lifecycle = lifecycle)
-    }?.collectAsLazyPagingItems()
-    val news2 = remember(flow2, lifecycle) {
-        flow2?.flowWithLifecycle(lifecycle = lifecycle)
-    }?.collectAsLazyPagingItems()
 
     NavHost(navController, startDestination = Navigation.HOME) {
 
@@ -320,10 +302,13 @@ fun NavGraph(
         //公告
         composable(Navigation.TOOL_NEWS) {
             viewModel.fabMainIcon.postValue(MainIconType.BACK)
+            val scrollState0 = rememberLazyListState()
+            val scrollState1 = rememberLazyListState()
+            val scrollState2 = rememberLazyListState()
             NewsList(
-                news0,
-                news1,
-                news2,
+                scrollState0,
+                scrollState1,
+                scrollState2,
                 actions.toNewsDetail
             )
         }
@@ -338,6 +323,7 @@ fun NavGraph(
             )
         ) {
             viewModel.fabMainIcon.postValue(MainIconType.BACK)
+
             val arguments = requireNotNull(it.arguments)
             NewsDetail(arguments.getString(Navigation.TOOL_NEWS_KEY) ?: "")
         }
